HACK: extcon: Add dummy USB extcon driver
Right now USB on MSM8916 relies on a functional extcon device.
Usually, you can craft one using an USB-ID GPIO or a MUIC extcon device,
but sometimes there is nothing useful available for use.

Ideally this should point to a device that provides the actual USB status
(usually the charger chip), but for initial bringup it's helpful to have
an extcon device that just forces USB gadget mode on permanently.